fix unrelated failed tests and reorganize Cypress commands
