+++ webapp/cypress/support/index.ts
@@ -1,4 +1,5 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import './commands'
+import './api_commands'
+import './ui_commands'
